# [cmd] Update fdtest to use the directfd driver and raw IO