The `fdtest` needed a long overdue update to work properly with TLVC and the `directfd` driver. Most of the old code (using direct BIOS calls) is now commented out: Access to raw IO allows a utility to do physical IO which is what this utility needs to do its testing.

Actually, the `dd` utility can do most of what `fdtest` does - with the help of the external `time` program. However, if a 'skip-length` - blocks to skip between each operation, is added, all kinds of new possibilities opens up for this program.

The timer needs to be fixed to report higher time resolution than 1 second.